websearch/pyproject.toml
[project]
name = "websearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.5.0",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
]
